# swagger.yaml
openapi: 3.0.0
info:

  title: e-commerce api

  contact: {}

  version: '1.0'
servers:

- url: https://islam-h-e-commerce-api.herokuapp.com/api/v1

  variables: {}
paths:
...
  /orders/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The order ID.

    get:

      tags:

      - orders

      summary: get single order

      operationId: getsingleorder

      parameters: []

      responses:

        '200':

          description: ''

          headers: {}

      deprecated: false
...
security:

- {}
...